requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-telegram-bot[job-queue]>=21.0
